Refactor data-sql types file structure
