+++ packages/data-sql/src/types/clauses/select/column.ts
@@ -1,0 +1,4 @@
+export interface AbstractSqlQueryColumn {
+	table: string;
+	column: string;
+}
