+++ packages/data-sql/src/types/clauses/where.ts
@@ -1,0 +1,4 @@
+import type { AbstractSqlQueryConditionNode } from './where/conditions.js';
+import type { AbstractSqlQueryLogicalNode } from './where/logical.js';
+
+export type AbstractSqlQueryWhereNode = AbstractSqlQueryConditionNode | AbstractSqlQueryLogicalNode;
